BACKEND/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
paho-mqtt==1.6.1
apscheduler==3.10.4
requests==2.31.0
pillow==10.1.0
opencv-python==4.8.1.78
httpx==0.25.2
pytest==7.4.3
black==23.12.0
flake8==6.1.0
